chore: refactor product_detail repo, route and services

+++ repo/product_detail_repo.py
@@ -1,22 +1,38 @@
-from models.product_detail import ProductDetails
 from instance.database import db
+from models.product_detail import ProductDetails
 
 
-def get_details_by_product_id(product_id):
-    return ProductDetails.query.filter_by(product_id=product_id).first()
+class ProductDetailRepository:
+    @staticmethod
+    def get_all():
+        return ProductDetails.query.all()
 
+    @staticmethod
+    def get_by_id(detail_id):
+        return ProductDetails.query.get(detail_id)
 
-def get_detail_by_id(detail_id):
-    return ProductDetails.query.get(detail_id)
+    @staticmethod
+    def create(data):
+        new_detail = ProductDetails(**data)
+        db.session.add(new_detail)
+        db.session.commit()
+        return new_detail
 
+    @staticmethod
+    def update(detail_id, data):
+        detail = ProductDetails.query.get(detail_id)
+        if not detail:
+            return None
+        for key, value in data.items():
+            setattr(detail, key, value)
+        db.session.commit()
+        return detail
 
-def create_product_detail(detail):
-    db.session.add(detail)
-
-
-def update_product_detail():
-    pass  # SQLAlchemy tracks changes
-
-
-def delete_product_detail(detail):
-    db.session.delete(detail)
+    @staticmethod
+    def delete(detail_id):
+        detail = ProductDetails.query.get(detail_id)
+        if not detail:
+            return False
+        db.session.delete(detail)
+        db.session.commit()
+        return True
